Handle differently spoon and spring-boot

+++ run.sh
@@ -18,25 +18,34 @@ echo "Copying ${input_repo_path} into ${output_repo_path}"
 echo "cp -r ${input_repo_path} ${output_repo_path}"
 cp -r "$input_repo_path" "$output_repo_path"
 
+# Create the .jar of object-instantiation project
 echo "mvn clean package"
 mvn clean package
 
+# Execute the .jar on the analyzed project
 cd ./target || return
 echo "java -jar object-instantiation-1.0-SNAPSHOT-jar-with-dependencies.jar ${input_source_code} ${output_source_code}"
 java -jar object-instantiation-1.0-SNAPSHOT-jar-with-dependencies.jar "$input_source_code" "$output_source_code" "$input_repo_path"
 
-# Copy be.unamur.snail.register.SendUtils to the output folder
+# Copy be.unamur.snail.register package to the output folder
 echo "mkdir -p ${output_source_code}/be/unamur/snail/register"
 mkdir -p "${output_source_code}/be/unamur/snail/register"
 echo "cp -r ${REGISTER_PATH} ${output_source_code}/be/unamur/snail/register"
 cp -r "$REGISTER_PATH" "${output_source_code}/be/unamur/snail/register/"
 
-# Execute spring-boot tests with the transformation
+# Execute tests of the analyzed project within the transformed code
 cd "$output_repo_path" || return
-rm -rf .gradle/
-export JAVA_HOME=/usr/lib/jvm/java-19-openjdk-amd64
-echo "./gradlew clean spring-boot-project:spring-boot:test"
-./gradlew clean spring-boot-project:spring-boot:test --rerun-tasks
 
-#export JAVA_HOME=/usr/lib/jvm/java-17-openjdk-amd64
-#mvn clean test
+if [ "$PROJECT_NAME" == "spring-boot" ]; then
+  export JAVA_HOME=/usr/lib/jvm/java-19-openjdk-amd64
+  rm -rf .gradle/
+  echo "./gradlew clean spring-boot-project:spring-boot:test"
+  ./gradlew clean spring-boot-project:spring-boot:test --rerun-tasks
+fi
+
+if [ "$PROJECT_NAME" == "spoon" ]; then
+  export JAVA_HOME=/usr/lib/jvm/java-17-openjdk-amd64
+  mvn clean test
+fi
+
+# TODO handle other projects execution
